pyproject.toml
[tool.poetry]
name = "react-django"
version = "0.1.0"
description = ""
authors = ["inessochan <iness0chan@gmail.com>"]
readme = "README.md"
packages = [{include = "react_django"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
pillow = "^9.3.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
